% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghostR-package.R
\docType{package}
\name{ghostR-package}
\alias{ghostR}
\alias{ghostR-package}
\title{ghostR: An R Package for Sensitivity Analyses with Phantom Variables}
\description{
This package allows you to create phantom variables to assess temporal bias or confounding bias in SEM models that do not have a mean structure.
}
\author{
\strong{Maintainer}: Alexis Georgeson \email{georgeson.alexis@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
