% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phantSEM-package.R
\docType{package}
\name{phantSEM-package}
\alias{phantSEM}
\alias{phantSEM-package}
\title{phantSEM: An R Package for Sensitivity Analyses in Structural Equation Models with Phantom Variables}
\description{
This package allows you to create phantom variables to assess temporal bias or confounding bias in SEM models that do not have a mean structure.
}
\author{
\strong{Maintainer}: Alexis Georgeson \email{georgeson.alexis@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
