% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA_step1.R
\name{SA_step1}
\alias{SA_step1}
\title{Sensitivity Analysis Function Step 1}
\usage{
SA_step1(lavoutput, mod_obs, mod_phant)
}
\arguments{
\item{lavoutput}{The lavaan output object output from lavaan functions sem() or lavaan() when fitting your observed model.}

\item{mod_obs}{A lavaan syntax for the observed model.}

\item{mod_phant}{A lavaan syntax for the phantom variable model.}
}
\value{
a list containing the names of all phantom covariance parameters
SA_step1()
}
\description{
This function generates the phantom variables and the names for their covariance parameters that will be used in SA_step2().
}
