% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA_step2.R
\name{SA_step2}
\alias{SA_step2}
\title{Step 2 of sensitivty analysis function}
\usage{
SA_step2(fixed_names, ref_names, varynames = NULL, testvals = NULL, step1)
}
\arguments{
\item{fixed_names}{this is a vector of the covariance parameter names that will be fixed to a single value}

\item{varynames}{a list or list of lists containing the names of the covariances that will be varied. If you wish to constrain certain parameters to be equal, you will need to put the names of these parameters in the same list and this argument will be a list of lists.}

\item{testvals}{a list or list of lists of values to try for each parameter.  If you are constraining certain parameters to be equal, you will put these in a list. See example.
SA_step2()}

\item{refnames}{this is a vector containing either single values or the names of other known parameters that you will fix each parameter in fixed_names to.}
}
\description{
This function is used to determine which phantom covariances will be fixed and which will be varied as part of the sensitivity analysis.
}
