% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA_step3.R
\name{SA_step3}
\alias{SA_step3}
\title{Step 3 of sensitivity analysis function}
\usage{
SA_step3(step2, n)
}
\arguments{
\item{step2}{the object returned from SA_step2.}

\item{n}{the sample size.
SA_step3()}
}
\description{
This function computes the parameter estimates in your phantom model defined in step 1 for the different values provided.
}
