% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA_step2.R
\name{SA_step2}
\alias{SA_step2}
\title{Step 2 of sensitivity analysis function}
\usage{
SA_step2(phantom_assignment, step1)
}
\arguments{
\item{phantom_assignment}{A list of all phantom parameter names (copied from \code{SA_step1()} output) which assigns them to be equal to ONE of the following: 1) an observed parameter name, 2) a single numeric value, 3) a sequence of values, or 4) another phantom variable that has been set equal to 1-3.}

\item{step1}{The output object created in \code{SA_step1()}}
}
\value{
A list containing test covariance matrices that the phantom model will be fit to.
}
\description{
\code{SA_step2()} is used to assign values to the phantom covariances. There are three options for assigning values to the phantom covariances: 1. Fix phantom covariance to a numeric value (i.e., 0 or 1), 2. Fix phantom covariance to be equal to another covariance, or 3. Test different values for the phantom covariance.
}
