% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA_step2.R
\name{SA_step2}
\alias{SA_step2}
\title{Step 2 of sensitivity analysis function}
\usage{
SA_step2(
  fixed_names,
  fixed_values,
  test_names = NULL,
  test_values = NULL,
  step1
)
}
\arguments{
\item{fixed_names}{this is a vector of the covariance parameter names that will be fixed to a single value}

\item{fixed_values}{A vector containing either single values or the names of other known parameters that you will fix each parameter in fixed_names to.}

\item{test_names}{A list of vectors containing the names of the covariances that will be varied. If you wish to constrain certain parameters to be equal, you will need to put the names of these parameters in the same vector.}

\item{test_values}{A list of seq() to try for each parameter.  If you are constraining certain parameters to be equal, you will put these in a list. See example.}

\item{step1}{The object created in \code{SA_step1()}}
}
\value{
A list containing test covariance matrices that the phantom model will be fit to.
}
\description{
\code{SA_step2()} is used to assign values to the phantom covariances. There are three options for assigning values to the phantom covariances: 1. Fix phantom covariance to a numeric value (i.e., 0 or 1), 2. Fix phantom covariance to be equal to another covariance, or 3. Test different values for the phantom covariance.
}
